Move ShiftEditCommand to command package
