+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -16,7 +16,7 @@ import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.ShiftEditCommand;
+import seedu.address.logic.commands.ShiftEditCommand;
 
 /**
  * Parses user input.
